Literally a complete re-write.
Switched to Eto.Forms for cross-platform support, which means I
completely re-wrote the UI
Added support for HookGen
Fixed a lot of small bugs and also re-did the patching system so
Partiality only patches when it needs to.
Mods and Patches are only different to mod developers and internally
now, there's no difference to users themselves, they put it all in the
"Mods" folder.
Disabling/enabling mods now does something.